requirements.txt

aiohttp
bs4
pillow
pydantic
fastapi
python-dateutil
